+++ rules/plugin-jsdoc.js
@@ -2,16 +2,23 @@
 
 module.exports = {
 	rules: {
+		'jsdoc/check-alignment': 'error',
 		'jsdoc/check-examples': 'error',
+		'jsdoc/check-indentation': 'error',
 		'jsdoc/check-param-names': 'error',
+		'jsdoc/check-syntax': 'error',
 		'jsdoc/check-tag-names': 'error',
 		'jsdoc/check-types': 'error',
+		'jsdoc/implements-on-classes': 'error',
 		'jsdoc/newline-after-description': 'error',
+		'jsdoc/match-description': 'error',
 		'jsdoc/no-undefined-types': 'off',
+		'jsdoc/no-types': 'off',
 		'jsdoc/require-description': 'off',
 		'jsdoc/require-description-complete-sentence': 'error',
 		'jsdoc/require-example': 'off',
 		'jsdoc/require-hyphen-before-param-description': 'error',
+		'jsdoc/require-jsdoc': 'off',
 		'jsdoc/require-param': 'error',
 		'jsdoc/require-param-description': 'error',
 		'jsdoc/require-param-name': 'error',